requirements.txt
flask
flask-sqlalchemy
pandas
python-dotenv
Pillow
pdfplumber
PyMuPDF
google-genai
tenacity
scikit-learn
numpy<2
sentence-transformers
